Explicitly exclude all .git directories from deploy
Explicitly excluding .git directories wasn't thought to be necessary
since the deploy script was extended to automatically archive Git
repositories. However, this only works for .git directories at the
root of the directory being deployed, assuming that directory
is a Git repository. This change ensures that .git directories
in subdirectories are also excluded from deployment.