Do not run mysql as root when starting/stopping MAMP
Apache should be run as root because it is running on port 80, which
macOS normally reserves for the system. However, MySQL is still running
on its default MAMP port (8889), which is not reserved by the system.
Therefore, the mysql command should not be run as root, especially since
strange things happened when I tried.