Revert "Extend virtualenv auto-detection to .venv directory"
This reverts commit 9d35a22e8127eeed5436762d16df067c639411a2.

The uv team is right that the virtualenv name should be a user setting,
not a project setting. It does not matter what my virtualenv directory
is called as long as it is the same for all projects. Therefore, the
next commit will set the UV_PROJECT_ENVIRONMENT variable to set
.virtualenv as my preferred virtualenv directory name.